requirements-dev.txt
black==22.3.0
Django==4.0.4
djangorestframework==3.13.1
ipython==8.3.0
isort==5.10.1
psycopg2-binary
pylama==8.3.8
python-dotenv==0.20.0
django-simple-history>=3.4.0,<3.5.0
dj-rest-auth>=5.0.1,<5.0.2
drf-spectacular>=0.26.5,<0.26.6
channels>=4.0.0,<4.0.1
daphne>=4.0.0,<4.0.1
channels-redis>=4.1.0,<4.2.0
selenium>=4.14.0,<4.15.0